backups/20250930_223609_pre_cleanup/src/requirements.txt
# Production Requirements
# Core FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis and caching
redis==5.0.1
slowapi==0.1.9

# Security and authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring and logging
sentry-sdk[fastapi]==1.38.0
prometheus-client==0.19.0
psutil==6.1.1

# Configuration
pydantic[email]==2.5.0
python-dotenv==1.0.0

# AI and ML dependencies
torch==2.1.0
transformers==4.35.2
ollama==0.1.7
pydantic-ai==0.0.14
# git+https://github.com/apple/ml-aim.git

# Web crawling and content processing
crawl4ai==0.7.4
beautifulsoup4==4.12.2
lxml==5.3.0
PyPDF2==3.0.1

# TTS Dependencies
edge-tts==6.1.9
torchaudio==2.1.0
chatterbox-tts==0.1.0
openai-whisper==20231117

# MLX (Apple Silicon)
mlx==0.8.1
mlx-lm==0.8.0

# Utilities
httpx==0.27.2
aiofiles==24.1.0
python-dateutil==2.8.2

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
